+++ web/src/locales/en-GB.ts
@@ -496,6 +496,7 @@ export default {
   Search: 'Search',
   'Search this library': 'Search this library',
   'See what would go': 'See what would go',
+  'See which groups you belong to': 'See which groups you belong to',
   'See your email address': 'See your email address',
   'See your name and username': 'See your name and username',
   'See your page': 'See your page',